data_ingestion/retriever_agent/requirements.txt
# API and Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0

# AI and ML
langchain>=0.1.0
openai>=1.12.0
InstructorEmbedding>=0.1.0
faiss-cpu>=1.7.0
sentence-transformers>=2.0.0
numpy>=1.21.0
scikit-learn>=0.24.0

# Utilities
pydantic>=1.8.0
loguru>=0.7.2
redis>=4.0.0
fastapi-limiter>=0.1.5
aioredis>=2.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0 